# [REVIEW] `Dataframe.sort_index` optimizations