@isVoid 2021-09-16
> consolidate df interface and series interface altogether as part of #9038 ?

This is a special case because we might want to avoid `Index.sort_index`.